requirements.txt
blinker==1.9.0
CacheControl==0.14.2
cachelib==0.13.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
firebase==4.0.1
firebase-admin==6.6.0
Flask==3.1.0
Flask-Session==0.8.0
fonttools==4.55.8
gcloud==0.18.3
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jws==0.1.3
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
msgpack==1.1.0
msgspec==0.19.0
numpy==2.2.2
oauth2client==4.1.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
pyparsing==3.2.1
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==2.0.1
pytz==2025.1
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
setuptools==75.8.0
six==1.17.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
Werkzeug==3.1.3
gunicorn==21.2.0